Move logging code into reused module
